fix upload image when avt, background is null in db

+++ src/main/java/com/example/socialnetwork/domain/service/CommentServiceImpl.java
@@ -188,8 +188,10 @@ public class CommentServiceImpl implements CommentServicePort {
 
         if(isDelete){
             String image = HandleFile.loadFileImage(images, storageServicePort, 1);
-            if(currentComment.getImage() != null){
+            if(currentComment.getImage() != null && !currentComment.getImage().isEmpty()) {
                 s3ServicePort.deleteFile(HandleFile.getFilePath(currentComment.getImage()));
+                currentComment.setImage(image);
+            }else {
                 currentComment.setImage(image);
             }
         }
